+++ unit_tests/others/CMakeLists.txt
@@ -1,4 +1,4 @@
-set(NAME unit_tests)
+set(NAME unit_tests_others)
 
 set(SOURCES
   adcs/AdcsCoordinatorTest.cpp    
@@ -12,7 +12,6 @@ set(SOURCES
   FileSystem/FileTest.cpp
   base/ReaderTest.cpp
   base/WriterTest.cpp
-  base/RapidTest.cpp
   base/OnLeaveTest.cpp
   base/RedundancyTest.cpp
   os/TimeoutTest.cpp
@@ -22,11 +21,6 @@ set(SOURCES
   time/SmartWait.cpp
   eigen/eigenTest.cpp  
   ErrorCounterTest.cpp    
-  state/PersistentStateTest.cpp
-  state/PersistentTimeStateTest.cpp
-  state/AntennaConfigTest.cpp
-  obc/ObcStateTest.cpp  
-  
 )
 
 add_executable(${NAME} ${SOURCES})
@@ -48,20 +42,17 @@ target_asm_listing(${NAME})
 target_compile_options(${NAME} PRIVATE "-fexceptions")
 
 target_link_libraries(${NAME}
+    yaffs
     m
     platform
     spi
     emlib
     logger
-    yaffs
     storage
     comm
     telecommunication
     fs    
-    base    
-    rapidcheck
-    rapidcheck_gtest
-    rapidcheck_gmock
+    base       
     emdrv
     time
     gsl
@@ -69,8 +60,7 @@ target_link_libraries(${NAME}
     eigen
     exp_fibo
     adcs
-    obc_telecommands
-    obc_state
+    obc_telecommands    
     eps
     rtc
     error_counter
@@ -80,7 +70,7 @@ target_link_libraries(${NAME}
 
 set (EXEC_OBJ ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${NAME})
 
-add_custom_target(unit_tests.run
+add_custom_target(${NAME}.run
   COMMAND ${QEMU} -board generic -mcu ${QEMU_MCU} -nographic -monitor null -image ${EXEC_OBJ} -semihosting-config "arg=tests,arg=--gtest_output=xml:${OUTPUT_PATH}/unit-tests.xml"
 
   DEPENDS ${NAME}
