+++ unit_tests/others/ErrorCounterTest.cpp
