+++ unit_tests/others/Logger/LoggerTest.cpp
